chore(IT Wallet): [SIW-2016] Remove remote feature flag from linking options (#6706)
## Short description
This PR removes the remote feature flag from the IT Wallet linking
options to allow routes to be configured properly at the app startup.

## List of changes proposed in this pull request
- Remove `isItwEnabledSelector` from `useItwLinkingOptions`
- Added `disabled` prop to `OperationResultScreenContent` actions props
- Added `isItwEnabledSelector` check on continuation CTA on every screen
that is exposed by a deep link

## How to test
Check that linking options for IT Wallet feature is working properly.
Disable the ITW feature flag and check that screens that are exposed by
a deep link have all continuation CTAs disabled.