feat: [IOPLT-959] Implements the Offline check loop feature (#6734)
> [!warning]
> This PR depends on pagopa/io-backend#1184

## Short description
This PR represents a first implementation of an infinite checker through
app session of its own connectivity status.
The Code is based on `@react-native-community/netinfo` library, checking
the connectivity on a specific endpoint exposed by our backend
infrastructure. The checker runs in a specific loop handled by a saga
that starts at the app opening such as the one reloading the app remote
configuration. At any check it writes up the result of the check on the
redux store in order to be read by any component of the app itself.

## List of changes proposed in this pull request
- implemented saga and store feature for connectivity check.

## How to test
Open the app and check the saga invoke on the determined time loop.