fix: [IOPLT-996] Enable quick dev access in the `IdentificationLockScreen` with screen reader enabled (#6769)
## Short description
This PR allows quick developer access in the `IdentificationLockScreen`
when the screen reader is enabled. Previously you were forced to enter
each single digit of the predefined unlock code (`defaultPin`), making
the developer experience even more miserable.

## List of changes proposed in this pull request
- Remove props that override accessibility features in the dev-only view

## How to test
Run the app in the local environment with VoiceOver/Talkback enabled.
You should be able to access the button with the unlock icon with the
screen reader.