fix: [IOPID-2484] Fix `MIXPANEL_SET_ENABLED` event tracking in the onBoarding flow (#6766)
## Short description
This PR replaces the wrong `flow` property set in the
`MIXPANEL_SET_ENABLED` event in the onBoarding flow with the correct
one.

## How to test
Perform a fresh install of the app and, once in the
`OnboardingShareDataScreen`, consent to data sharing. The `flow`
property associated with the `MIXPANEL_SET_ENABLED` event should be set
to `onBoarding`.

Co-authored-by: Fabio Bombardi <16268789+shadowsheep1@users.noreply.github.com>