chore: [IOBP-1231] Removed total amount from IdPayCard if reward type is `EXPENSE` (#6771)
## Short description
This PR removes the total amount label from the IdPayCard shown in the
wallet home screen if the IdPay initiative is of the `EXPENSE` type.

## List of changes proposed in this pull request
- Updated the IDPay openAPI version;
- Added a conditional render inside the `IdPayCard` to show the total
amount;

## How to test
With the dev-server started, onboard a new IDPay EXPENSE type and check
that you don't see the total amount inside the wallet home page.