chore(IT Wallet): [SIW-2055] Update Mixpanel events ITW_DATA_SHARE and ITW_DATA_SHARE_ACCEPTED for async driving license flow (#6757)
## Short description
This PR moves the logic of the two Mixpanel events `ITW_DATA_SHARE` and
`ITW_DATA_SHARE_ACCEPTED` from `ItwIssuanceCredentialTrustIssuerScreen`
to the XState machine of credential issuance
`itwCredentialIssuanceMachine`.

## List of changes proposed in this pull request
- Created a new type, `ItwWalletDataShare`, which includes the
`credential type` and the request `phase`. The `phase` is only included
if the credential type is a Driving License.
- Added two new actions in `itwCredentialIssuanceMachine`:
`trackCredentialIssuingDataShare` and
`trackCredentialIssuingDataShareAccepted`
- Introduced the `trackDataShareEvent` function, used by both actions to
generate the tracking data. This function determines whether to include
the request phase based on the credential type, whether there is an
ongoing request for a Driving License, and whether the flow originates
from `ItwIssuanceCredentialAsyncContinuationScreen` (indicating an
asynchronous flow) or from `ItwCredentialOnboardingSection`

## How to test

1. Request a credential other than the driving license and check that
the two events are sent correctly without a phase
2. Request the driving license and check that the two events have an
`initial_request` as phase
3. Upon receiving the message, proceed with the asynchronous flow and
check that the phase is async_continuation

---------

Co-authored-by: Federico Mastrini <federicomastrini93@gmail.com>