chore: [IOBP-1243,IOBP-1244,IOBP-1241] a11y improvements on payment method details card preview (#6779)
## Short description
This PR improves the `a11y` behavior from the payment method detail
screen by adding a more detailed summary label when focusing on the
payment method card.

## List of changes proposed in this pull request
- Wrapped the `PayPalLogo` component inside a `View` from the
`PaymentCard` component to be accessible with a description;
- Added the `accessibilityRole` to the view that wraps the payment card
detail from the Payment method details;

## How to test
- With the SR enabled, navigate the payment method details screen and
check that `PayPal` is announced correctly when focusing on the payment
card and that `Validity` is announced correctly if the payment is a
card.

---------

Co-authored-by: Emanuele Dall'Ara <71103219+LeleDallas@users.noreply.github.com>