fix: [IOPLT-989] Add proper configuration of `threshold` value to the `ForceScrollDownView` component (#6787)
> [!caution]
> This PR depends on:
> * https://github.com/pagopa/io-app-design-system/pull/423

## Short description
This PR updates the configuration of the `ForceScrollDownView` component
by setting the `threshold` prop to the correct value. The previous
values were hard-coded and not linked to the measures of the
`FooterActions` component.

## List of changes proposed in this pull request
- Update `io-app-design-system` to `5.1.0`
- Add the new `footerActions` prop instead of inserting `FooterActions`
as a React child. This change ensures that the `threshold` is properly
configured with the correct and updated value of the height of the
action block
- Add two new DS pages for testing purposes

### Preview

| Stress test | DS screens |
|--------|--------|
| <video
src="https://github.com/user-attachments/assets/ab8fe1c2-c710-4d61-91e1-a06df4e02c59"
/> | <video
src="https://github.com/user-attachments/assets/4e7ef4d3-5164-4a70-9297-3e018cc9a1d2"
/> |


## How to test
Go to the one of the screens edited in this PR and check the behavior in
different device screen sizes