chore(IT Wallet): [SIW-2036] Prevent user interactions with components requiring internet access (#6782)
> [!WARNING]
> Depends on #6778
 
## Short description
This PR implements ways to prevent user interactions with components
that require an internet access, such as screen with remote content or
CTAs that trigger network requests.

### Preview


https://github.com/user-attachments/assets/f5dac7f3-bdad-40b5-9ecc-c4be3022bc68

## List of changes proposed in this pull request
- Added `OfflineFailureComponent` and `OfflineFailureScreen`
- Implementation of `useOfflineGuard` utility hook that allow to wrap a
function and change its behavior based on connectivity status
- Wrapped the trustmark button and revocation CTAs with
`useOfflineGuard` hook

## How to test
Go to Settings → Documenti su IO → Enable offline access .
With the app online → check if there aren't regressions
With the app offline → with a valid Wallet Instance, check if you are
able to see your wallet with all the credentials, try to navigate to
screens or press buttons that requires internet access.

---------

Co-authored-by: Alice Di Rico <alice.dirico@pagopa.it>
Co-authored-by: Alice Di Rico <83651704+Ladirico@users.noreply.github.com>
Co-authored-by: Fabio Bombardi <16268789+shadowsheep1@users.noreply.github.com>