feat: [IOCOM-2100] Common FIMS start API (#6740)
⚠️ This PR depends on
[939](https://github.com/pagopa/io-services-metadata/pull/939) ⚠️
⚠️ This PR depends on
[468](https://github.com/pagopa/io-dev-api-server/pull/468) ⚠️

## Short description
This PR centralise the starting of the FIMS authentication flow.

## List of changes proposed in this pull request
There are now three hooks that allow to properly start the FIMS flow
from a react-native component
- `useFIMSFromServiceId`: when you know the service Id ahead of time and
the service details are already loaded into memory (e.g, from message
and service details)
- `useFIMSAuthenticationFlow`: when you do not know the service Id ahead
of time but will have it upon authentication flow invokation (e.g., a
bonus of IdPay)
- `useFIMSRemoteServiceConfiguration`: when the service data are stored
in the CDN and you know the configuration id associated with such
configuration (e.g., from the wallet driving license). Be aware that
calling this hook with an invalid configuration Id will not start the
authentication flow later on

All these hooks return an authentication callback that starts the FIMS
authentication flow upon invocation.
 
## How to test
Using the io-dev-api-server, check that all existing FIMS authentication
flows are working as expected.

---------

Co-authored-by: Martino Cesari Tomba <60693085+forrest57@users.noreply.github.com>