chore: [IOAPPX-494] Update node binary script detection on iOS (#6840)
## Short description
This PR updates the node binary detection script on iOS according to the
one provided directly from React Native for the current version of the
app.

## List of changes proposed in this pull request
- Update the `project.pbxproj` with the new script which automatically
detects and create a `.xcode.env.local` file which contains the result
of `which node`;
- Update `README.md` instructions to build the app accordingly.

## How to test
Try to run a iOS build from Xcode without manually changing the `node`
path.

---------

Co-authored-by: Andrea <andrea.piai@pagopa.it>