feat: [IOCOM-2235,IOCOM-2240] Push token absence tracking and fix for missing token on iOS (#6815)
## Short description
This PR tracks the push token presence and fixes a case where the token
was not requested.

## List of changes proposed in this pull request
- Added NOTIFICATION_TOKEN property to profile properties
- On iOS, the push notification permission has to be requested on every
application run. Failing to do so results in a missing token generation.
In a use-case where the application has already received the system
(push notification) permission, the related saga was not requesting them
and the token was not generated
- Fixed the Android push-notification dev configuration (not regenerated
after the canary project creation). All of its data can be made public
since it does not contain any secret nor it is used in production
- Updated the RootContainer configuration so the store can be passed
down to the push configuration utilities, without having to relying on
the exported store.

## How to test
Using the io-dev-api-server check that:
- on the iOS simulator / device, the token is generated on every
application run, when the saga that asks for permissions runs
- on Android, check that the token is generated upon application
initialization (i.e., even before the permissions saga has run)

In both cases, check that the `NOTIFICATION_TOKEN` property in the
mixpanel's profile properties is properly generated and sent

---------

Co-authored-by: Martino Cesari Tomba <60693085+forrest57@users.noreply.github.com>