feat: [IOCOM-2282] New copy for denied calendar permission popup (#6845)
## Short description
This PR changes the copy of the popup that is displayed when the user ha
denied the calendar access permission, starting from the message details
screen.

|Android|iOS|
|-|-|

|![Android](https://github.com/user-attachments/assets/ad1c4012-88a3-4069-b4ed-65dd667116fe)|![iOS](https://github.com/user-attachments/assets/7f85afcb-a5d2-47e4-8e1e-cd8bf88e35af)|

## List of changes proposed in this pull request
- Removed popup title
- Changed popup description
- The related popup in the profile settings has been left unchanged

## How to test
Using the io-dev-api server, set
`paymentInvalidAfterDueDateWithValidDueDateCount` and
`paymentWithValidDueDateCount` to more than one, open such message, try
to add the reminder to the calendar and deny the permission. Try again
(on Android, two more times) and check that the popup text is as shown
in the screenshots above.