feat(IT Wallet): [SIW-2048] Add new device re-enable itw banner (#6833)
## Short description
This PR updates the itw onboarding banner, adding a new version that is
enabled if the user already has a Wallet Instance active but not locally
on the device. This is useful if the user changes device, so he can
understand he must re-activate "Documenti su IO"

## List of changes proposed in this pull request
- Update `io-react-native-wallet` to `0.28.0` refactoring also the
`uuid` usage. There is also a jest fix for `transformIgnorePatterns`
- Add a new property inside the `preferences` called
`isWalletInstanceRemotelyActive` that indicates if the user has a
remotely active Wallet Instance but is not locally active.
- Add a new saga to fetch the current wallet instance status without the
`id` but based on the Fiscal Code of the user
- Add some tests
- Add translations for the new banner

## How to test
Given the fact, that I updated the `uuid` library using the new one
provided by `io-react-native-wallet`, an entire ITW flow test needs to
be done.
Another test could be activating ITW on a device, then uninstalling the
app or activating ITW on another device. This should trigger the new
re-activate modal.

> [!NOTE]  
> If you revoke the WI, the re-activate banner will not be enabled. It
will only be enabled if you have remotely the WI active but not locally
on the device.


Screenshots:

![IMAGE 2025-03-19
17:05:50](https://github.com/user-attachments/assets/d7c55e48-f385-4298-951f-0b3465dc4b19)
![IMAGE 2025-03-19
17:05:57](https://github.com/user-attachments/assets/f36a4947-8a42-401a-bc1b-f7eabe28f879)