chore(IT Wallet): [SIW-2174] Remove lifecycle state (#6846)
## Short description
This PR removes the `lifecycle` reducer and the associated
`ItwLifecycleState`, as it was just duplicating a state that can be
inferred from the wallet content. Eliminating it reduces an unnecessary
abstraction layer and simplifies the wallet logic.

## List of changes proposed in this pull request
- Removed `lifecycle` reducer, `ItwLifecycleState` enum and lifecyle
actions
- Updated tests and imports

## How to test
Test eID and credential issuance flows and verify that no regressions
are introduced.