chore(IT Wallet): [SIW-2198]  Improve serializeFailureReason for Mixpanel tracking   (#6870)
## Short description
This PR improves and fixes the `reason` property in the
`ITW_ID_REQUEST_UNEXPECTED_FAILURE `event to ensure it is always
serialized as an object, not a plain string.

## List of changes proposed in this pull request
- Create an object with `code: UNEXPECTED`, agreed upon with the
Mixpanel team, and `errorDescription: failure.reason.message`

## How to test
I was able to replicate the situation described in the support tickets
using Charles Proxy, where the error was Network request failed. In
detail, the error appears as `{"reason": [TypeError: Network request
failed], "type": "UNEXPECTED"}`. Now the reason is on Mixpanel should be

`{"code":"UNEXPECTED","errorDescription":"Network request failed"}`