chore(IT Wallet): [SIW-2164] Update PID machine for L3 Wallet (#6875)
## Short description
This PR updates the eID issuance flow to support the issuance of an L3
PID for users who have enabled the L3 mode via feature flag.

## List of changes proposed in this pull request
- Bumped `@pagopa/io-react-native-wallet` to `0.28.2`
- Added `isItwL3Enabled` parameter to the `startAuthFlow` function,
which allows to omit `idpHint` parameter in the ITW's `/credential`
request
- Updated ITW's `preferences` reducer to persist some data after a
`itwLifecycleStoresReset` action

## How to test
> [!IMPORTANT]
> **It is currently not possible to test the full L3 PID issuance with
the issuer**. Tests should be based on the presence or absence of the
`idphint` parameter in the `/authorize` request during eID issuance.

Check that everything works as expected by requesting a new eID and your
credentials.

Steps to verify:
1. Open the app and go to:  
   `Settings > Developers > Documenti su IO > IT Wallet (L3)`
2. Enable the **"Enable L3 wallet"** switch
3. Request a new eID using CIE+PIN
4. Inspect the `/authorize` request:
- When L3 is enabled, ensure the `idphint` parameter is **not present**