feat: [IOBP-1563] Add remote banner specific for payment method PSP screen (#6883)
## ⚠️ This PR depends on
https://github.com/pagopa/io-services-metadata/pull/958

## Short description
This PR adds a dismissable remote banner specific for a payment method
from the PSP selection screen;

## List of changes proposed in this pull request
- A `WalletPaymentPspBanner` has been added to manage the remote
configuration of banners displayed on the PSP selection screen. The
banner's visibility can be controlled based on the payment method name,
and its dismissal is valid for the entire payment session. Once the
payment session ends, the banner may reappear if the remote
configuration is enabled.
- An orchestration action `paymentMethodPspBannerClose` has been created
to store the payment method names where a dismissal has occurred into a
`Set`.

## How to test
- Test this PR by running the io-dev-api server with the following PR:
https://github.com/pagopa/io-dev-api-server/pull/477.
- Run `yarn generate`.
- Start a payment flow in the development environment and verify that
only the Bancomat Pay payment method displays the banner (you can modify
which payment method shows the banner by editing the pspBanner attribute
in the dev-server).
- Verify that after dismissing the banner for the first time within the
same payment session, the banner is no longer displayed.
- Check that if you initiate a new payment flow in a different session,
the banner is displayed again for that payment method.

## Preview


https://github.com/user-attachments/assets/4e29b84a-f4cf-4207-b0e4-3f564c2dba95

---------

Co-authored-by: Emanuele Dall'Ara <71103219+LeleDallas@users.noreply.github.com>