fix: [IOBP-1552] Saved payment methods profile property with right value (#6890)
## Short description
This PR fixes the value of the `SAVED_PAYMENT_METHODS` profile property.
Prior to this fix, the property was set to a value that was always
undefined when launching the app for the first time, causing incorrect
spikes in Mixpanel with uncorrect data.

## List of changes proposed in this pull request
- Mapped the correct value to the profile property;
- Added a dispatcher to update the profile property by increasing or
decreasing the saved payment methods count based on whether the user is
adding or removing a payment method.

## How to test
Check that when launching the app for the first time, as well as when
adding or removing a payment method, the `SAVED_PAYMENT_METHODS` profile
property is correctly set.