feat(IT Wallet): [SIW-2207] Remote presentation, inactive IT-Wallet failure screen (#6900)
## Short description
This PR modifies the copy for the `ItwRemoteFailureScreen` in the case
of an inactive wallet.
Show an alert using `useItwDismissalDialog`, in case you want to exit
from the remote presentation flow
## List of changes proposed in this pull request
- Updated the copy for the `ItwRemoteFailureScreen` in the wallet
inactive case
- Modified `useItwDismissalDialog` to accept a custom body message.

## How to test
Without IT-Wallet scan a valid QR code for remote presentation, check
that the screen displayed is correct and clicking get “IT-Wallet” takes
you to the activation flow. Clicking “Not now” should display the alert
asking for confirmation to close the process or not.





https://github.com/user-attachments/assets/7fbe24b0-7e7d-499f-bf48-58a2feb0ea3e

Co-authored-by: Federico Mastrini <federicomastrini93@gmail.com>