fix: [IOCOM-2224] Proper tracking of the push notification opening analytics event (#6878)
## Short description
This PR fixes the tracking of the push notification opening analytics
event (`NOTIFICATIONS_MESSAGE_TAP`), which was tracked twice.

## List of changes proposed in this pull request
- New mixpanel tracking queue that allows to enqueue tracking events
when mixpanel is not initialized yet (and the user has opted in)
- Push opening tracking removed from the message opening and moved at
push opening time
- Mixpanel file refactored to hide the private instance
- Push notifications analytics events guarded against an uninitialized
mixpanel with opted-out user
- Missing tests for refactored code

## How to test
Test that the event is generated on both Android and iOS when:
- a notification is received in foreground
- a notification is received in background, paused app
- a notification is received in background, stopped/killed app

Do the same tests with an opted-out user, making sure that the event is
not tracked.

User login is a must.