chore(IT Wallet): [SIW-2055] Update iPatente cta url with saved driving license number (#6886)
## Short description
This PR updates the URL for the iPatente service to allow direct access
to the details of a specific driving license document and to view the
points balance associated with it.

## List of changes proposed in this pull request
- Updated the cta url to include the document number, if present

## How to test
1. Go to driving license details
2. Click on the iPatente service "Vai al saldo punti patente"
3. Verify that you are redirected to the specific points balance of your
driver's license

> [!NOTE]  
> It is not possible to check the full flow currently, a change is
needed on the web app that manages services such as IPatente. To check
however you can retrieve the generated url and paste it into your phone
browser, you should be redirected to the specific license points balance