fix(IT Wallet): [SIW-2290] Disable wallet offline access when no credentials are stored (#6910)
## Short description
This PR introduces updates to improve the handling of IT Wallet offline
access and refactors related selectors and logic for better
maintainability. The most critical changes involve consolidating
multiple selectors into a new unified selector,
`itwOfflineAccessAvailableSelector`, and updating various components,
sagas, and tests to use this new selector.

## List of changes proposed in this pull request
- Added `itwOfflineAccessAvailableSelector` which returns if the wallet
can be accessed offline by checking if the offline access is enabled and
the wallet is in a valid state with credentials
- Updated `ingress` features by replacing old selectors with the new one
- Updated tests

## How to test
Check that the wallet is accessible offline only with a valid wallet
with at least one credential stored

---------

Co-authored-by: Alice Di Rico <83651704+Ladirico@users.noreply.github.com>