fix(IT Wallet): [SIW-2339] Disable status alert for offline mini-app (#6950)
## Short description
This PR addresses a bug which causes the status alerts to be displayed
within the offline mini app, causing a double alert to appear

## List of changes proposed in this pull request
- Added `offlineAccessReason` check within `useStatusAlertProps` hook

## How to test
Ensure that status alerts works as expected across all app features and
no regression were introduced.
With the device offline, open the wallet screen and check that appears a
single alert which informs of the offline status.