feat(IT Wallet): [SIW-2211,SIW-2273] Add IT Wallet upgrade banner (#6929)
> [!IMPORTANT]
> Depends on https://github.com/pagopa/io-app/pull/6921 

## Short description
This PR add the IT Wallet's upgrade banner to the wallet screen.

## List of changes proposed in this pull request
- Added `ItwUpgradeBanner` component
- Added `ItwUpgradeBanner` to the wallet screen
- Refactored wallet's banners components

## How to test
Make sure that no regressions have been introduced and that discovery
banners are displayed correctly.
Go to **Settings > Playground > Documenti su IO**, enable L3 feature by
switching the **Enable L3 wallet** toggle.
Go to the wallet screen, check that the banner is correctly rendered,
both with the wallet active and inactive.

## Demo


https://github.com/user-attachments/assets/b41d1741-fd72-4db7-8586-54bba6b8b232

---------

Co-authored-by: RiccardoMolinari95 <98079356+RiccardoMolinari95@users.noreply.github.com>