feat(IT Wallet): [SIW-2301] Track exception on sentry (#6913)
## Short description
This PR adds exception handling in critical points of the code within
the IT Wallet flow. If exceptions are thrown, they are sent to Sentry.

## List of changes proposed in this pull request
- Add `beforeSendHandler` function in Sentry config. This function
checks if the event should be sent based on the `isSendRequired`
property in the event's context. If `isSendRequired` exists, the event
will always be sent to Sentry. **For events that do not require sending,
a sampling rate of 20% will be applied**
- The sample rate has been set to 1 in Sentry config
- Add `itwSendExceptionToSentry` function, that captures exceptions and
sends them to Sentry with additional context and tags. It includes an
optional parameter, `sendRequired`, to enforce sending the exception
regardless of sampling logic.

## How to test
Trigger an exception by setting the `sendRequired` parameter to true
when calling the `itwSendExceptionToSentry` function. After triggering
the exception, check Sentry to verify that the exception is displayed
correctly, ensuring that it bypasses any sampling logic.