docs: [IOPID-2990] startup saga analysis  (#6977)
## Short description
As an A&I team, we analyzed the startup saga and tried to figure out
where we need to work to resolve some technical debt and to begin to
unbundle some logic.
We also jumped at the opportunity to make a diagram describing what
flows are executed within the startup saga.