feat(IT Wallet): [SIW-2347] Enable attestation generation with fallback over TEE (#6995)
## Short description
This PR adds a TEE fallback when the attestation generation with
StrongBox enabled fails.

## List of changes proposed in this pull request
- Updated `io-react-native-integrity` to v0.3.2 to implement this
feature

## How to test
With an Android device, try to activate ITW and check that no
regressions are present. The preferred test would be to have a device
that supports Strongbox but has issues with it, so although there is
support for Strongbox, a fallback to TEE would be made. This case
occurred on some Motorola and Xiaomi devices.

> [!NOTE]
> Testable from QA; however, as previously written, this test will be
effective on devices with a Strongbox issue.