feat(IT Wallet): [SIW-2444] Add `useOfflineToastGuard` when opening FIMS from MDL details (#7016)
This PR adds the `useOfflineToastGuard` hook to protect the FIMS flow
with the offline toast.

## Short description
Include a summary of the changes.

## List of changes proposed in this pull request
- Added `useOfflineToastGuard` hook to protect the FIMS flow with the
offline toast.

## How to test
When using the full app in offline mode or the mini-app in offline mode,
navigate to the MDL details screen and press the "Vai al saldo punti
patente" CTA. The offline toast should pop up.

Co-authored-by: Federico Mastrini <federicomastrini93@gmail.com>