fix(IT Wallet): [SIW-2469] Disable gesture navigation in the final presentation screen (#7033)
## Short description
This PR disables gesture navigation in `ItwRemoteAuthResponseScreen`.
Currently it is possible to swipe back to the claims disclosure screen,
breaking the flow.

## List of changes proposed in this pull request
- Call `useItwDisableGestureNavigation` hook

## How to test
Complete the remote presentation (same-device and cross-device) on iOS.
In the last screen (`ITW_REMOTE_AUTH_RESPONSE`) try to swipe back: it
should not be possible to navigate to the previous screen.