chore: [IOBP-1723] Disable pagoPA UAT toggle if app is in local env (#7040)
## Short description
This PR introduces changes to improve the handling of the pagoPA test
environment in the developer mode section of the settings. It supports
detecting local app builds and disabling the pagoPA test environment in
such cases, along with corresponding updates to localization files and
UI components.

## List of changes proposed in this pull request
- Modified the `DeveloperTestEnvironmentSection` to use `isLocalEnv` to
conditionally set the description and disable the toggle for the pagoPA
test environment.

## How to test
Open the developer settings page with an app built with the .env.local
variables and check that the `pagoPA` switch is disabled.

## Preview
The following preview shows both the case when the app is built with a
local .env and another with the production .env


https://github.com/user-attachments/assets/64574470-2abc-4c72-9a63-5a975199d426